// src/react-app/lib/finance-types.ts
export interface GlobalBank {
  codigo: string;
  nome: string;
  logoUrl?: string;
}